# openapi-spec.yml
openapi: '3.0.2'
info:
  title: CyberDAS API
  version: '0.6'
  contact:
    email: api@cyberdas.net
servers:
  - url: https://api.cyberdas.net/v1
  - url: https://api.cyberdas.net/next
...
paths:
...
  /queues/{queue}/slots/{id}/reserve:

    parameters:
      - $ref: '#/components/parameters/queue'
      - $ref: '#/components/parameters/id'
      - $ref: '#/components/parameters/emailNext'

    post:
      summary: Занять выбранный слот в выбранной очереди
      tags:
        - Очереди
      security:
      - cookieAuth: []
        csrfToken: []

      requestBody:
        $ref: '#/components/requestBodies/QuickAuth'

      responses:
        '201':
          description: Успешная запись. На вашу почту выслана копия талона.
            Если ваши планы изменились, то вы можете отменить запись, пройдя
            по ссылке из письма.
        '401':
          $ref: '#/components/responses/UnauthenticatedError'
        '403':
          description: Этот слот уже занят\истёк или Вы уже записались в эту очередь
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          $ref: '#/components/responses/UnexpectedError'
    
    delete:
      summary: Отменить бронирование слота
      tags:
        - Очереди
      security:
        # Нужно либо cookieAuth, либо adminAuth
        - cookieAuth: []
        - adminAuth: []

      parameters:
        - $ref: '#/components/parameters/quickAuthToken'

      responses:
        '204':
          description: Запись удалена
        '401':
          $ref: '#/components/responses/UnauthenticatedError'
        '403':
          $ref: '#/components/responses/UnauthorizedError'
        default:
          $ref: '#/components/responses/UnexpectedError'
components: 
  schemas: 
...
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
  responses:
    UnexpectedError:
      description: Непредвиденная ошибка
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    UnauthenticatedError:
      description: Неизвестный пользователь
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    UnauthorizedError:
      description: Отсутствует право доступа
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  parameters:
    queue:
      name: queue
      in: path
      required: true
      description: Имя очереди
      schema:
        type: string
        example: "living"
    id:
      name: id
      in: path
      required: true
      description: ID запрашиваемого объекта
      schema:
        type: integer
        example: 23
    quickAuthToken:
      name: token
      in: query
      required: false
      allowEmptyValue: false
      description: Подписанный токен, позволяющий аутентифицироваться на 
        некоторых эндпоинтах без процедуры логина и регистрации
      schema:
        type: string
        format: byte
    emailNext:
      name: next
      in: query
      required: false
      allowEmptyValue: false
      description: Параметр, который будучи указан, перенаправляет пользователя
        из письма на {frontend}/{next}, а не обратно на API
      schema:
        type: string
  requestBodies:
...
    QuickAuth:
      description: JSON-объект, содержащий адрес почты для логина или все 
        данные, необходимые для регистрации. Позволяет пропустить регистрацию 
        и использовать некоторые защищенные эндпоинты без неё.
      required: false
      content:
        'application/json':
          schema:
            type: object
            properties:
              email:
                type: string
              name:
                type: string
              surname:
                type: string
              patronymic:
                type: string
              faculty_id:
                type: integer
              course:
                type: integer
            required:
              - email
...
  securitySchemes:

    cookieAuth:
      type: apiKey
      in: cookie
      name: SESSIONID

    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token

    ottAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

    adminAuth:
      type: apiKey
      in: cookie
      name: ADMINID


# Для совершения всех действий требуется активная сессия, если не указано иное
security:
  - cookieAuth: []